requirements.txt
pandas==2.3.0
yfinance==0.2.64
PyYAML==6.0.2
matplotlib==3.10.3
duckduckgo_search==8.0.4
scikit-learn==1.7.0
tensorflow==2.19.0
numpy==2.1.3
streamlit==1.46.1
openai==1.93.0
openpyxl==3.1.5
python-dotenv